refactor: add agenda and details folder in components
