+++ androidApp/src/main/java/com/example/tasky/android/agenda/components/agenda/AgendaFloatingActionButton.kt
@@ -1,4 +1,4 @@
-package com.example.tasky.android.agenda.components
+package com.example.tasky.android.agenda.components.agenda
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.shape.RoundedCornerShape
