+++ androidApp/src/main/java/com/example/tasky/android/agenda/components/agenda/AgendaTopBar.kt
@@ -1,4 +1,4 @@
-package com.example.tasky.android.agenda.components
+package com.example.tasky.android.agenda.components.agenda
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
