+++ androidApp/src/main/java/com/example/tasky/android/agenda/components/details/DetailsDateTimeSection.kt
@@ -1,4 +1,4 @@
-package com.example.tasky.android.agenda.components
+package com.example.tasky.android.agenda.components.details
 
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
