+++ androidApp/src/main/java/com/example/tasky/android/agenda/components/details/DetailsDescSection.kt
@@ -1,8 +1,6 @@
-package com.example.tasky.android.agenda.components
+package com.example.tasky.android.agenda.components.details
 
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.width
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.outlined.KeyboardArrowRight
 import androidx.compose.material3.Icon
@@ -11,30 +9,22 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.tasky.android.R
 import com.example.tasky.android.theme.Black
 import com.example.tasky.android.theme.MyApplicationTheme
 
 @Composable
-fun DetailsTitleSection(
-    title: String,
+fun DetailsDescSection(
+    desc: String,
     isEdit: Boolean,
     modifier: Modifier = Modifier,
 ) {
     Row(modifier = modifier, verticalAlignment = Alignment.CenterVertically) {
-        Icon(
-            painter = painterResource(R.drawable.outline_circle_24),
-            contentDescription = null,
-        )
-        Spacer(Modifier.width(8.dp))
         Text(
-            title,
-            style = typography.displaySmall,
-            lineHeight = 25.sp,
+            desc,
+            style = typography.bodySmall,
+            lineHeight = 15.sp,
             color = Black,
             modifier = Modifier.weight(1f),
         )
@@ -45,10 +35,14 @@ fun DetailsTitleSection(
     }
 }
 
-@Preview
+@Preview(showBackground = true, backgroundColor = 0xFFFFFFFF)
 @Composable
-private fun DetailsTitleSectionPreview() {
+private fun DetailsDescSectionPreview() {
     MyApplicationTheme {
-        DetailsTitleSection("Project X", true)
+        DetailsDescSection(
+            "Weekly plan\n" +
+                "Role distribution",
+            true,
+        )
     }
 }
