+++ androidApp/src/main/java/com/example/tasky/android/agenda/screen/AgendaDetailsScreen.kt
@@ -27,18 +27,18 @@ import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import com.example.tasky.android.R
-import com.example.tasky.android.agenda.components.AgendaEditText
-import com.example.tasky.android.agenda.components.AgendaEditTextType
-import com.example.tasky.android.agenda.components.DetailsAttendeeSection
-import com.example.tasky.android.agenda.components.DetailsAttendeeSectionTabOption
-import com.example.tasky.android.agenda.components.DetailsDateTimeSection
-import com.example.tasky.android.agenda.components.DetailsDeleteSection
-import com.example.tasky.android.agenda.components.DetailsDescSection
-import com.example.tasky.android.agenda.components.DetailsHeaderSection
-import com.example.tasky.android.agenda.components.DetailsPhotoSection
-import com.example.tasky.android.agenda.components.DetailsRemindAtSection
-import com.example.tasky.android.agenda.components.DetailsTitleSection
-import com.example.tasky.android.agenda.components.DetailsTopBar
+import com.example.tasky.android.agenda.components.details.AgendaEditText
+import com.example.tasky.android.agenda.components.details.AgendaEditTextType
+import com.example.tasky.android.agenda.components.details.DetailsAttendeeSection
+import com.example.tasky.android.agenda.components.details.DetailsAttendeeSectionTabOption
+import com.example.tasky.android.agenda.components.details.DetailsDateTimeSection
+import com.example.tasky.android.agenda.components.details.DetailsDeleteSection
+import com.example.tasky.android.agenda.components.details.DetailsDescSection
+import com.example.tasky.android.agenda.components.details.DetailsHeaderSection
+import com.example.tasky.android.agenda.components.details.DetailsPhotoSection
+import com.example.tasky.android.agenda.components.details.DetailsRemindAtSection
+import com.example.tasky.android.agenda.components.details.DetailsTitleSection
+import com.example.tasky.android.agenda.components.details.DetailsTopBar
 import com.example.tasky.android.agenda.viewmodel.AgendaDetailsViewModel
 import com.example.tasky.android.theme.Black
 import com.example.tasky.android.theme.Light
